third_party/WebKit/LayoutTests/external/wpt/webrtc/simplecall-expected.txt

CONSOLE WARNING: line 114: getUserMedia() no longer works on insecure origins. To use this feature, you should consider switching your application to a secure origin, such as HTTPS. See https://goo.gl/rStTGz for more details.
This is a testharness.js-based test.
FAIL Can set up a basic WebRTC call. assert_unreached: WebRTC called error callback for getUserMedia Reached unreachable code
Harness: the test ran to completion.
